refactor: split UnifiedAIAssistant (745→71) em hook + componentes
Extrai a lógica (estado, gravação/transcrição, imagem, análise, sugestões)
para o hook useUnifiedAIAssistant e quebra o JSX em componentes
presentacionais controlados (AIInputArea, AIResultPanel, IdentifiedCustomerCard,
IdentifiedItemsLists, SuggestionsList). Tipos públicos movidos para types.ts e
re-exportados pelo arquivo principal (preserva imports de useUnifiedOrder/
UnifiedOrder). Helpers puros (fmt/getToolName/formatDuration) em helpers.ts.
Comportamento preservado 1:1; +31 testes. Warning exhaustive-deps é baseline
(idêntico ao original).
